# s2: Add arm64 decompression