--- a/.travis.yml
+++ b/.travis.yml
@@ -4,8 +4,11 @@ os:
   - linux
   - osx
 
+arch:
+  - amd64
+  - arm64
+
 go:
-  - 1.13.x
   - 1.14.x
   - 1.15.x
   - 1.16.x