--- a/.travis.yml
+++ b/.travis.yml
@@ -23,7 +23,7 @@ install:
 
 script:
  - diff <(gofmt -d .) <(printf "")
- - go test -cpu=2 ./s2
+ - go test -v -cpu=2 ./s2
 # - go test -cpu=2 -tags=noasm ./...
 # - CGO_ENABLED=1 go test -cpu=1,4 -short -race ./...
  - go build github.com/klauspost/compress/s2/cmd/s2c && go build github.com/klauspost/compress/s2/cmd/s2d && s2c s2c && s2d s2c.s2 && rm s2c && rm s2d && rm s2c.s2