--- a/s2/decode_asm.go
+++ b/s2/decode_asm.go
@@ -3,6 +3,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+// +build amd64 arm64
 // +build !appengine
 // +build gc
 // +build !noasm