Adapted from Snappy equiv. Minor optimizations along the way.
```
enwik9 (stream):
Decompressing. 426854614 -> 1000000000 [234.27%]; 7.618s, 125.2MB/s
Decompressing. 426854614 -> 1000000000 [234.27%]; 4.258s, 224.0MB/s

Blocks:
benchmark                     old ns/op     new ns/op     delta
BenchmarkTwainDecode1e1-4     69.5          63.0          -9.27%
BenchmarkTwainDecode1e2-4     184           153           -16.61%
BenchmarkTwainDecode1e3-4     852           522           -38.74%
BenchmarkTwainDecode1e4-4     30220         14678         -51.43%
BenchmarkTwainDecode1e5-4     570195        231880        -59.33%
BenchmarkTwainDecode1e6-4     3349753       1920724       -42.66%

benchmark                     old MB/s     new MB/s     speedup
BenchmarkTwainDecode1e1-4     143.97       158.69       1.10x
BenchmarkTwainDecode1e2-4     544.75       653.07       1.20x
BenchmarkTwainDecode1e3-4     1174.00      1916.39      1.63x
BenchmarkTwainDecode1e4-4     330.90       681.27       2.06x
BenchmarkTwainDecode1e5-4     175.38       431.26       2.46x
BenchmarkTwainDecode1e6-4     298.53       520.64       1.74x
```